requirements.txt
requests
pillow
beautifulsoup4
playwright
